WPCS formatter: Minor change to the version number comment format.
Was: `// x.x.x`
Now: `// WP x.x.x.`

To eliminate as much noise as possible in the comparison, I've started adding the version nr comments to the array in the WPCS sniff as well.

For the comments to pass the `WordPress-Docs` sniffing, they need to end in a period.
What with just having a number + period looking a bit off, adding the `WP` makes is a little more phrase like.